Add new test: Checking to see if  cloneing a specific git tag by using add  works
