--- test/util/fallbackSB07.js
@@ -1,183 +1,0 @@
-module.exports = fallbackDependancySandBox
-
-function fallbackDependancySandBox (appDir) {
-  const fs = require('fs')
-  const path = require('path')
-  const testSrc = path.resolve(__dirname, '../../test')
-  const { execSync } = require('child_process')
-  const repoList = ['repo19', 'repo20', 'repo21']
-
-  if (!appDir) {
-    let processEnv
-    if (fs.existsSync(path.join(process.cwd(), 'node_modules')) === false) {
-      processEnv = process.cwd()
-    } else {
-      processEnv = undefined
-    }
-    appDir = processEnv
-  }
-
-  // Checks if the repos folder exist, if not it creates it
-  try {
-    if (!fs.existsSync(`${testSrc}/repos`)) {
-      // creat the repos folder in the test folder
-      fs.mkdirSync(`${testSrc}/repos`, err => {
-        if (err) {
-          console.error(err)
-        }
-        // file written successfully
-      })
-    }
-    // Checks if the clone folder exist, if not it creates it
-    if (!fs.existsSync(`${testSrc}/clones`)) {
-      // creat the clones folder in the test folder
-      fs.mkdirSync(`${testSrc}/clones`, err => {
-        if (err) {
-          console.error(err)
-        }
-        // file written successfully
-      })
-    }
-
-    const repo19Package = {
-      devDependencies: {
-        'fallback-dependencies': '../../../'
-      },
-      fallbackDependencies: {
-        dir: 'lib',
-        reposFile: 'reposFile.json'
-      },
-      scripts: {
-        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
-      }
-    }
-    const repo19PackageLocked = {
-      name: 'repo19',
-      lockfileVersion: 3,
-      requires: true,
-      packages: {
-        '': {
-          hasInstallScript: true,
-          devDependencies: {
-            'fallback-dependencies': '../../..'
-          }
-        },
-        '../../..': {
-          version: '0.1.0',
-          dev: true,
-          license: 'CC-BY-19.0',
-          devDependencies: {}
-        },
-        'node_modules/fallback-dependencies': {
-          resolved: '../../..',
-          link: true
-        }
-      }
-    }
-    const repo19FileData = {
-      'fallback-deps-test-repo-20': [
-        'https://github.com/rooseveltframework/generator-roosevelt.git -b 0.21.7', 'https://githu.com/rooseveltframework/roosevelt.git -b 0.21.12'
-      ]
-    }
-    const repo20Package = {
-      devDependencies: {
-        'fallback-dependencies': '../../../../../../'
-      },
-      fallbackDependencies: {
-        dir: 'lib',
-        repos: {
-          'fallback-deps-test-repo-21': '../../../../../repos/repo21'
-        }
-      },
-      scripts: {
-        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
-      }
-    }
-    const repo20PackageLocked = {
-      name: 'repo19',
-      lockfileVersion: 3,
-      requires: true,
-      packages: {
-        '': {
-          hasInstallScript: true,
-          devDependencies: {
-            'fallback-dependencies': '../../../../../'
-          }
-        },
-        '../../../../..': {
-          version: '0.1.0',
-          dev: true,
-          license: 'CC-BY-19.0',
-          devDependencies: {}
-        },
-        'node_modules/fallback-dependencies': {
-          resolved: '../../../../..',
-          link: true
-        }
-      }
-    }
-    const repo21Package = {}
-    const repo21PackageLocked = {
-      name: 'repo21',
-      lockfileVersion: 3,
-      requires: true,
-      packages: {}
-    }
-
-    const packageList = [[repo19Package, repo19PackageLocked], [repo20Package, repo20PackageLocked], [repo21Package, repo21PackageLocked]]
-    for (const id in repoList) {
-      if (!fs.existsSync(`${testSrc}/repos/${repoList[id]}/`)) {
-        fs.mkdirSync(`${testSrc}/repos/${repoList[id]}/`, err => {
-          if (err) {
-            console.error(err)
-          }
-          // file written successfully
-        })
-      }
-      // Change directory path run git command
-      execSync('git --bare init', {
-        stdio: 'pipe', // hide output from git
-        cwd: path.resolve(`${testSrc}/repos/${repoList[id]}`, '') // where we're cloning the repo to
-      })
-      // Change directory path run git command
-      execSync(`git clone ${testSrc}/repos/${repoList[id]}`, {
-        stdio: 'pipe', // hide output from git
-        cwd: path.resolve(`${testSrc}/clones`, '') // where we're cloning the repo to
-      })
-      if (repoList[id] === 'repo19') {
-        fs.writeFileSync(`${testSrc}/clones/repo19/reposFile.json`, JSON.stringify(repo19FileData))
-      }
-      // Change directory path
-      process.chdir(`${testSrc}/clones/${repoList[id]}`)
-      // Create the package.json and package-lock.json file in the ./clones/..
-      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package.json`, JSON.stringify(packageList[id][0]))
-      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package-lock.json`, JSON.stringify(packageList[id][1]))
-
-      // Run git command to push package and package-lock files
-      execSync('git add .', {
-        stdio: 'pipe', // hide output from git
-        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
-      })
-      execSync('git commit -m "commit"', {
-        stdio: 'pipe', // hide output from git
-        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
-      })
-      execSync('git push', {
-        stdio: 'pipe', // hide output from git
-        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
-      })
-    }
-    // Run git command to push package and package-lock files
-    const we = execSync('npm ci', {
-      stdio: 'pipe', // hide output from git
-      cwd: path.resolve(`${testSrc}/clones/repo19`, '') // where we're cloning the repo to
-    })
-    const wa = execSync('npm ci', {
-      stdio: 'pipe', // hide output from git
-      cwd: path.resolve(`${testSrc}/clones/repo19`, '') // where we're cloning the repo to
-    })
-
-    // console.log(we.toString())
-    // console.log(wa.toString())
-  } catch {}
-}
